pyproject.toml
[project]
name = "employee-goal-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "mostlyai-mock>=0.1.7",
    "openai>=1.86.0",
    "pandas>=2.3.0",
]
